Check for minimal supported Gradle version.
