+++ plugin/src/test/kotlin/org/jlleitschuh/gradle/ktlint/BuildCacheTest.kt
@@ -3,6 +3,7 @@ package org.jlleitschuh.gradle.ktlint
 import org.assertj.core.api.Assertions.assertThat
 import org.gradle.testkit.runner.GradleRunner
 import org.gradle.testkit.runner.TaskOutcome
+import org.jlleitschuh.gradle.ktlint.KtlintBasePlugin.Companion.LOWEST_SUPPORTED_GRADLE_VERSION
 import org.junit.jupiter.api.Test
 import java.io.File
